Move API calls to separate module
