+++ cli/src/rdwatch_cli/api/__init__.py
@@ -1,0 +1,4 @@
+from .image import image
+from .movie import movie
+
+__all__ = ["image", "movie"]
