+++ cli/src/rdwatch_cli/api/image.py
@@ -2,16 +2,12 @@ import asyncio
 import io
 from dataclasses import dataclass
 from datetime import datetime
-from pathlib import Path
 
 import aiohttp
-import click
 import mercantile
 from PIL import Image
 
 from rdwatch_cli.exceptions import ImageNotFound, ServerError
-from rdwatch_cli.utils import get_http_client
-from rdwatch_cli.validators import validate_bbox, validate_timestamp
 
 
 @dataclass
@@ -34,16 +30,11 @@ async def fetch_tile(
         else f"/api/satellite-image/tile/{tile.z}/{tile.x}/{tile.y}.webp"
     )
 
-    params = (
-        {"timestamp": time.isoformat()}
-        if worldview
-        else {
-            "constellation": "S2",
-            "timestamp": time.isoformat(),
-            "spectrum": "visual",
-            "level": "2A",
-        }
-    )
+    params = {"timestamp": datetime.isoformat(time)}
+    if not worldview:
+        params["constellation"] = "S2"
+        params["spectrum"] = "visual"
+        params["level"] = "2A"
 
     async with client.get(url, params=params) as resp:
         buffer = await resp.read()
@@ -56,9 +47,9 @@ async def fetch_tile(
 
 
 async def image(
+    client: aiohttp.ClientSession,
     bbox: tuple[float, float, float, float],
     time: datetime,
-    client: aiohttp.ClientSession | None = None,
     worldview: bool = False,
 ) -> Image.Image:
     # The max zoom will be different for lower resolution imagery
@@ -67,21 +58,12 @@ async def image(
     zoom = 18 if worldview else 14
 
     webp_tiles: list[WebpTile]
-    if client is None:
-        async with get_http_client() as client:
-            webp_tiles = await asyncio.gather(
-                *(
-                    fetch_tile(client, time, tile, worldview=worldview)
-                    for tile in mercantile.tiles(*bbox, zoom)
-                )
-            )
-    else:
-        webp_tiles = await asyncio.gather(
-            *(
-                fetch_tile(client, time, tile, worldview=worldview)
-                for tile in mercantile.tiles(*bbox, zoom)
-            )
+    webp_tiles = await asyncio.gather(
+        *(
+            fetch_tile(client, time, tile, worldview=worldview)
+            for tile in mercantile.tiles(*bbox, zoom)
         )
+    )
 
     xmin = min(t.tile.x for t in webp_tiles)
     xmax = max(t.tile.x for t in webp_tiles)
@@ -112,24 +94,3 @@ async def image(
     crop = merged.crop((crop_left, crop_top, crop_right, crop_bottom))
 
     return crop
-
-
-@click.command()
-@click.option("--output", type=Path)
-@click.option("--bbox", nargs=4, type=click.UNPROCESSED, callback=validate_bbox)
-@click.option("--time", type=click.UNPROCESSED, callback=validate_timestamp)
-@click.option("--worldview", type=bool)
-def image_cmd(
-    output: Path,
-    bbox: tuple[float, float, float, float],
-    time: datetime,
-    worldview: bool = False,
-) -> None:
-    """Retrieve an image for a bounding box at a specific time"""
-    try:
-        img = asyncio.run(image(bbox, time, worldview=worldview))
-        img.save(output)
-    except ImageNotFound:
-        raise click.ClickException("No image found for the given bounding box and time")
-    except ServerError:
-        raise click.ClickException("Server down (try again later)")
